Further dependency management changes
This may not be the direction we go, but make everything internally
consistent for now. Remove the setup target in favor of only an init
target that works the same as other services. Remove the stray dev
dependencies in pyproject.toml and fix tox to install from the pinned
hashes.