Rework the secret models
The secret model inheritance was confusing and required some copies
that were unnecessary for the type system. We were also resolving
the top-level secret condition during environment loading, but not
resolving conditions on copy and generate until secret resolution,
but there's no reason not to resolve all the conditions at the same
time.

Restructure the models to have Conditional versions of not only the
secret but also the copy and generate rules, and resolve all of the
conditions during configuration loading. Rely on the type system to
catch any use of the condition after configuration loading.