Fix some secret description bugs
Ensure that secrets can be listed for the IDF environments and for
minikube. The T&S environments are not yet working because they
want to take the Slack alert webhook from mobu and mobu isn't
enabled for those environments. This will be fixed eventually by
enabling mobu without any monkeys.