Rework the secret generator models
Move the code to generate a secret into a method on the secret
generator rules model, since there is no business logic there.
Add the understanding that certain generation types require a
source key directly to the model so that Pydantic will do model
checking.

The second change unfortunately makes the model hierarchy more
complex because of conditions. Move the definition of the condition
attribute to a mixin to avoid repeating so much code, and add a
couple of type aliases to simplify things. This unfortunately
still requires an isinstance check when generating secrets, but it
keeps the YAML describing a secret fairly simple.