Support secret sync with no static secrets source
If no static secret source was provided but all of the required
static secrets are already in Vault, successfully sync and retain
the secrets in Vault. This is required to support the USDF
environments, where static secrets are maintained directly in Vault,
and in general will be more convenient since it will mean that
1Password secrets don't have to be provided when one is just syncing
new generated secrets.