Environment confiuguration cleanup
Rename the environment repoURL setting to repoUrl to match our
camel-case convention for Helm chart values (and what CamelCaseModel
will expect for Pydantic). Add quoting to all of the application
templates that were underquoted for consistency. Remove the
strimzi-registry-operator application, since we now use it only as
a subchart for other applications and it was disabled for every
environment. Add values.yaml to the values file list for
alert-stream-broker, since it seems to be in use.