Detect location of the Phalanx configuration
Add a flag to every phalanx CLI command that uses the Phalanx
configuration specifying its location, and default to searching
upwards from the current directory to find the top of the
configuration tree (by looking for the applications and
environments directories).

Add a wrapper function around Click command-line testing for the
test suite that adds the path to the test configuration directory
and flips the default for catching exceptions to False.