Improve help output from CLI
The generic help command required some changes to support subcommands
and fix the usage line of the help output. The top level command was
still referring to Gafaelfawr rather than Phalanx. Also flesh out the
help text for all of the subcommands and document the required
VAULT_TOKEN settings consistently.