Add API and schema documentation
Add an internals page that is available only via the top bar from
the pull-down menu, since most people will never need to see it.
Document the JSON schemas there (currently only one for secrets.yaml
but more will come later) and add internal API documentation.

This is of arguable usefulness, but sometimes it can be nice to
browse the API documentation when working on new code, and it has
the nice benefit of checking all of the docstrings more thoroughly
than a linter does.