Add schema checking for environments
Now that we can fully model the schema for the values files for
environments, add a command to generate that schema, store the
current version in the schema directory, and enable schema checking
for those files.

Add a Python test that parses the full Phalanx configuration as if
we were about to generate documentation. This is redundant with
actually generating documentation, but the Python test suite is much
faster to run, so this catches problems earlier when iterating.