Update developer documentation
Go through the about and developer documentation and try to update
them to the current state of Phalanx. Flesh out hte developer
documentation and try to provide a clearer flow between the steps
needed to add a new application.

Start documenting the new secrets management approach, but add a lot
of warnings that it's not currently live. Rework the developer
documentation about secrets into a single page about defining
secrets that discusses secrets.yaml as well as the current
generate_secrets.py approach and adding a secret to 1Password.

Remove the chart overview page and include its contents where
appropriate in other pages. Also move documentation about creating
a chart from the add an application page to the one page about
writing charts.

Add a new empty Helm chart starter and recommend it for new
applications that aren't web services.