Simplify structure of environent config
Rather than using a structure for each application that has a set
of key/value pairs, of which only enabled is generally present or
used, instead have the applications section of an environment just
contain a key for every application and a value of true or false.
This simplifies the model and the configuration files. We don't want
to put per-application settings into the environment configuration,
so there's no need to leave space for it.

Fix the construction of the list of enabled applications for
telegraf and telegraf-ds. This creates a leading @ since argocd is
no longer special, so adjust the telegraf-ds ConfigMap definition
accordingly.