Use :download: for JSON schemas
Reference the JSON schemas with :download: to avoid linkcheck
issues and ensure they still work correctly in local builds of the
documentation. This isn't precisely correct since it makes a second
copy of the schemas, but given that I'm including this just for the
sake of completionism and very few people will ever bother following
these links, it seems good enough.

Thanks to Jonathan Sick for the suggestion.