Add onepassword-connect-dev application
This application will create the 1Password Connect API servers for
all development and test 1Password vaults for SQuaRE-run Phalanx
environments.

Because we need to create a separate subchart per 1Password vault
and dependencies with aliases can't be templated using values.yaml,
unfortunately we'll need to create two Phalanx applications, one
that collects the API servers for the development and test vaults
and runs on roundtable-dev, and one that collects the API servers
for the production vaults and runs on roundtable. This is the one
for development and test servers, starting with just IDF dev.