Add 1Password support for static secrets
Add new configuration options for environments that allow specifying
a 1Password Connect server and Vault title. If those options are set
for an environment, use that, via onepasswordconnectsdk, to retrieve
static secrets provided that a separate static secrets YAML file was
not given.

The layout used for 1Password is to have one 1Password item per
application, with one field (without section nesting) for each secret.