Move starter to option, constrain values
Rather than only doing runtime checking that the specified Helm
starter to phalanx application create is correct, add an enum for
the available starters and do the checking inside Click. Pass that
enum down to the other layers of the code, which eliminates the
need for an exception.

This isn't ideal, since it requires code changes when adding a new
starter, but I don't see a good alternative given that the Phalanx
configuration tree path is not known until runtime due to the -c
flag. Try to make this more robust by adding a test that will fail
if the enum gets out of sync with the starters that exist in the
configuration tree.

Also make web-service the default starter and move the starter to
an option, which can be omitted. This provides somewhat better help
output.