Allow applications to be disabled everywhere
Previously, after converting to the new models, applications had to
be enabled in at least one environment for their documentation to
build. Otherwise, there would be no Jinja context for the application
and the documentation build would fail with an obscure error message.

Fix this by building Jinja contexts for every application mentioned
in any environment values.yaml file. This allows new applications to
be added without being enabled, and doubles as a check to ensure there
are no typos or unknown applications mentioned in the environment
configurations.