Support encoding secrets in 1Password
Some of our secrets contain significant newlines, such as the AWS
credentials and pgpass file for Butler. 1Password unfortunately
does not allow embedded newlines in fields. The official solution
to this problem is to use files, but files can no longer be attached
to Server items and instead need to be separate Document items,
which breaks our entire 1Password layout.

Instead, support encoding the secret in base64 when storing it in
1Password and decoding it when retrieving it. This should allow
arbitrary values at the cost of some complexity and having to
explicitly tag which secrets are encoded in base64.