Add pull-secret handling
Restructure the static secrets format and template to move the
application secrets under an applications key and add a new
pull-secret top-level key with details about the pull secret.
This is in the form of a dict of registries to username and password
pairs.  The encoded .dockerconfigjson secret key is generated during
secret sync.

Add 1Password support for retrieving the pull secret. This 1Password
item uses one section per registry with fields username and password
specifying the HTTP Basic Auth credentials.

Change the phalanx secrets onepassword-secrets command to write a
YAML file in the same syntax as is used for static secrets. Not only
is this easier to compare, it makes it possible to download secrets
once from 1Password and then perform multiple audit or sync operations
without the 1Password credentials, if desired.