Centralize handling of optional files in prompt-proto-service.
Previously, configuring the prompt processing service to use an
in-container repo required pointing `instrument.calibRepo` to a local
path, setting the `centralRepoFile` flag to download the repo secret,
and explicitly mounting one into the other. Automating the third step
in the Helm chart instead of forcing the user to specify it as
configuration reduces room for human error.

A similar argument applies to the `cred_file_auth` flag. Since we don't
currently use this flag, the mount configuration for it is a bit of
a guess.