Update to Pydantic v2
Use alpha versions of Safir and documenteer for Pydantic v2 support.

Update for the API changes in Pydantic v2 and avoid all deprecated
methods. The most intrusive changes are around schema generation,
which is significantly different in the new version and requires
some additional class methods to enable the tricks we want to use.