Switch to Vault AppRole authentication for idfdev
Convert idfdev to use Vault AppRole authentication instead of a
token. Use the defaults for max token lifetime, since we're not
imposing any particular limits. Use vault-credentials as the name
of the secret instead of vault-secrets-operator so that we don't
clash with the default secret name for the service account.