Support nublado-lab-secret with new secret system
In the new secret management system, there is a single nublado
application secret, which contains both secrets for JupyterHub
itself and also secrets that are exposed to the user's lab. Add
support for this structure when generating nublado-lab-secret
in Kubernetes by using Vault Secrets Operator templating to extract
from the nublado application secret only those secrets that are
referenced by the lab configuration.