Update secrets documentation for new method
The old 1Password approach isn't supported any more and some of the
infrastructure has been turned off, so delete the old documentation.
We're not going to follow it again; if we have to, we'll make
temporary Vault changes until we can switch over.

Do an editing pass on the secrets management documentation and
simplify for having only one approach. Remove a bunch of warnings
that should hopefully now be irrelevant.