Better handling of missing 1Password secrets
Retrieve each 1Password item separately rather than using the bulk
API so that we can do proper error reporting of all of the missing
secrets we expected to be in 1Password but which weren't found.
Throw that as an exception up through sync, but, in audit, catch
that exception and properly report the missing secrets.