Add secrets config for more applications
Add configuration for the new secrets management system to obsloctap,
plot-navigator, and production-tools. All of these only used the old
Butler secret, so copy the secret from Nublado like we do for other
uses of the Butler secret. As with datalinker, add a configuration
option that determines whether to use the new per-application secret
or the old shared Butler secret, and default it to false.

Fix some helm-docs comments around the image tag.