Improve generated schemas
Add more annotations to the Pydantic models so that the generated
schemas contain more details. Mark the Vault URL as an HTTP URL for
Pydantic validation purposes. Set a default URL pointing to the
SQuaRE Vault server so that most environments don't need to set it,
and adjust the schema so that this will validate. Drop the
now-obsolete onepasswordUuid enviroment setting.