Add encoded secret warnings to static templates
If a secret is marked as base64-encoded in 1Password, add a warning
to the static secret template saying so, in case people are using
that as a reference to decide what to put into 1Password.

This required some refactoring to export the correct fields from
the static secrets model for a template. In the process, I fixed a
few places where we may have incorrectly synced an empty pull secret
when no pull secret was specified.