[DM-40944] Use my private docker image, dont use read only
First, we don't have a repo building sia images yet, that will come
after CADC gets things sqquared away.  I'm using my built image
off of Pat's repo until we can handle that.

Next, the default for images is to keep things readonly,
but tomcat doesn't like that with the rename war file,
which allows us to not have to do a bunch of proxy rewriting.