[DM-40944] Fix up uws database
There's so much going on here.  First off, don't tell it in the
URL that it is a database named uws, that is assumed, and won't
work if you put that in there.  It's hardcoded to the name uws.

So instead of the tap service uws database, which is behind,
and it will try to be upgraded, let's use the library/postgres
image, and the only thing I need to do is create the schema.

So the configmap has this config script and mounts it in the
postgres container in the spot where it runs the initial
scripts.