Support linting only changed applications
Add a --git flag to phalanx application lint-all that compares
against the origin/main branch and only lints application and
environment pairs that could have changed based on the changes
between those Git branches.

Pass a path argument to helm lint to make it easier to analyze
the test results.