Rework the syntax for phalanx application lint
Rather than taking the environment as an optional second argument,
allow multiple applications to be specified so that they can all be
linted at the same time, and move the environment to a flag. This
makes more sense since usually linting for all configured environments
is the desired behavior (it's very fast, so usually there's no reason
to save the time).