Remove $RUBIN_INSTRUMENT from prompt-proto-service values.
The $RUBIN_INSTRUMENT variable is defined in the Helm template
(Kubernetes config). Its use in the pipelines config is a holdover from
the pre-Helm setup, when it reduced the amount of code that needed to
be edited to switch instruments. With the Phalanx setup, all it does is
introduce a (conceptual) circular dependency from the values file back
to the template.