Adjust conditions on uploading docs
Now on pull requests, only upload the docs if content in the docs/
directory itself changes. This saves us CI time for changes that
incidentally change the docs; we still CI that the docs will build and
the docs themselves will update on the final merge to main.

In re-working the conditional, I've also found that we don't have a good
overal strategy for tags; we'd need to generally build the docs for any
tag. I've dropped the "tags" from the "on" clause. But if we decide to
start tagging Phalanx we'd want to revisit that.