Restrict minikube testing
Previously, we always ran minikube testing for changes to Phalanx
applications even if that application was not deployed on minikube.
That's pointless, wastes resources, and risks false positives.
Skip minikube testing unless the change affects one of the applications
deployed on minikube, and remove hips from that set since we don't
run any meaningful tests against it.