Remove nublado2
Retire the nublado2, cachemachine, and moneypenny applications in
favor of the nublado application. Move some documentation from those
applications to elsewhere in the Phalanx documentation.

CC-IN2P3 and RoE had not yet moved to Nublado v3, so write initial
configurations for them for Nublado v4. These are probably not fully
correct, but hopefully they can serve as a basis for future work.