Add new jira-data-proxy application
Developed at https://github.com/lsst-sqre/jira-data-proxy the
jira-data-proxy provides a read-only (only GET methods) proxy to the
Jira REST API. A user can access this proxy using their common
Gafaelfawr token and send GET requests to the Jira API through the
proxy, which uses a built-in LDAP credential.

This service is developed for Times Square reports and intended to be
deployed for Rubin staff at USDF.

Note: I needed to fix the name of vaultSecretsPathPrefix in the
Application template; this will have to be fixed in the Phalanx
templates going forward.