Specify butler repositories in Phalanx
Instead of using a file, specify Butler server repositories directly in the Phalanx configuration.  This will allow services to restart automatically when the list changes, and doesn't require us to hand-edit live configuration files out in the environment.

Shortly there also will not be a good place for the configuration files to live, since services using Butler will no longer have the S3 credentials to the buckets where these files are currently located.