Remove minikube documentation and scripts
The documentation for running minikube was out of date (it didn't
reflect the change to using app roles) and isn't tested, since we've
been using development environments instead. Remove it, at least for
now; if we want to bring it back, we should come up with a
well-tested replacement with a clear way to add and test specific
applications.