Add support for Gafaelfawr schema upgrades
Add a new configuration option to Gafaelfawr, config.updateSchema,
that if set runs gafaelfawr update-schema from a Helm pre-install
and pre-upgrade hook before installing the rest of the application.

This requires duplicating the ConfigMap and VualtSecret, since even
though Helm hooks have a mechanism for retaining resources that are
created in hooks, Argo CD wants to remove them again. Create a new
ServiceAccount as well, but don't use it yet since the Terraform
configuration is not yet merged, and therefore it is not yet
available.

The schema update job has to reintroduce some complex logic and a
separate image tag suffix for its Cloud SQL Auth Proxy so that it
will shut down properly when the schema update is complete. Eventually
we can eliminate this by using the Cloud SQL Auth Proxy Operator.