Add Prompt Processing service option to turn off calib caching.
The option prevents crashes when calibs are fed out of chronological
order to our validity workaround code. It should not be set for any
production run.

The option is not set in any of the instrument-specific values.yaml
files to minimize its exposure; this should make it easier to remove
once the underlying issue is fixed.