Add more probes for Gafaelfawr
Add a liveness problem for both the main Gafaelfawr deployment and
the Kubernetes operator. The former uses the new /health endpoint,
and the latter uses the Kopf-provided /healthz endpoint. If there
is more than one Gafaelfawr replica, switch the readiness probe for
the main Gafaelfawr deployment to use the same health check as the
liveness probe so that traffic will be routed away when it starts
failing.