Directly specify repositories for Butler server
Instead of reading a repository index file from the environment to configure Butler server, directly specify the list of repositories in an environment variable.

In an upcoming commit adding the server to USDF, we will need the list of repositories to be different than the usual index file for the environment.  It is also just better to have this configuration revision-controlled instead of off in some file on a random server.