Move Vault token duration to the chart
We previously were configuring the Vault token duration in the
secret, which makes absolutely no sense since its value is not a
secret. Move this configuration to the chart, setting the chart
default to the same length of time we were previously putting in
the secret.

Quote all of the values files for vault-secrets-operator for
consistency, and delete pointless environment variable settings
that are handled by the chart.

Update the vault-secrets-operator bootstrapping documentation for
the new secrets management system.