Add new write:files scope
Define a new write:files scope and assign it to all users on the
IDF deployments where we plan to run the WebDAV file server. Change
the scope used to control access to WebDAV from exec:notebook to
write:files, and add this to the list of scopes delegated to the
Portal.