Add phalanx environment install command
Port the installer/install.sh script to Python as the new phalanx
environment install command. Add a strongly worded confirmation
message before proceeding, which can be overrriden with a command
line flag for CI testing.

Expand the Helm storage layer and add Argo CD and Kubernetes
storage layers to support the installer. Use kubectl rather than
the Python Kubernetes libraries to manipulate Kubernetes objects,
since it's simpler and more straightforward for the type of actions
the installer needs to take.

Introduce a new way of passing around Vault credentials, and support
authenticating to Vault with an AppRole, at least in the context of
the installer. (The secrets and vault commands still require tokens.)

Remove the old installer and switch GitHub Actions CI over to the
new command.