Move TLS handling to Argo CD
We need to designate one application that sets the TLS configuration
in its index so that cert-manager will create a certificate. Follow
Roundtable and use Argo CD for this purpose rather than squareone,
since Argo CD is bootstrapped early and will thus ensure TLS is working
earlier in the process, and because technically squareone is only for
the RSP and may not be deployed in all Phalanx environments.