Handle the case of a nonexistent Vault path
Do not error out if the Vault path doesn't exist in secrets audit
and secrets sync, since otherwise we can't create the tree for a
new environment from scratch. Unfortunately, the Vault server
returns permission denied for nonexistent paths, so we can't
distinguish between this case and using the wrong Vault token until
we try to write.