Fix repeated environment installs
The second time one ran phalanx environment install for the same
environment with the same cluster and partial resources, it would
fail because it tried to create the vault-secrets-operator secret
and could not overwrite it. Change the installer to use kubectl
apply instead of kubectl create, similar to the old bash script,
which avoids this problem. As a bonus, this will avoid leaking the
Vault tokens to error output if the secret creation fails.