Cache authentication for Portal and Nublado
Cache authentication for up to five minutes, using a cache key that
invalidates the cache if the Cookie or Authorization headers have
changed, for Portal and the Nublado proxy ingress. We want to avoid
forcing Gafaelfawr to validate every HTTP request for interactive
web sites that may load a large number of ancillary resources.