Use null instead of the empty string in starters
null in YAML is a more accurate representation of an unset value
and, for applications that use part of the values file directly as
a configuration file for the application, will translate into the
more-useful None value in Python.

Fix the documentation comment for image.tag to specify a default.