Test for application/environment config issues
Make documentation generation robust against applications not being
configured for any environment. Add a test that will fail in this
case, and also test that an application is configured for every
environment for which it is enabled. This will catch various issues
people had during the bootcamp in a more obvious way.