Document time intervals for Nublado
Several times in the Nublado configuration related to culling were
given in number of seconds with no human-readable equivalent. Add
comments with the human-readable time.

Remove some cull override settings that should be safe to keep at
the defaults (we don't use named servers, and it should be fine to
remove the user on cull), and remove some obsolete settings that
should no longer be needed. Change defaults for unset settings to
null instead of the empty string.