Enable minimum token lifetime for Nublado
Use the Gafaelfawr functionality we added some time back and require
a minimum token lifetime when spawning a lab. This lifetime can be
explicitly configured or taken from the maxAge lab culler setting.

Change the culler defaults to match the settings we're using for
most environments so that the default maxAge is not longer than the
default Gafaelfawr token lifetime.