Convert base cluster to new secrets layout
Convert the base cluster to the new secrets layout with 1Password
secrets management. This required changes to the rubintv shared
chart to parameterize its secrets management and allow for the new
secrets layout.

Note that rubintv uses aws_credentials.txt and aws-credentials.txt
currently, which is probably a bug that the rubintv developers will
need to unwind.