Add additional datastore for dp02 raws
Add a new relative root that can be used to reference the DP02 raw image files.

Raw images for DP02 were stored in a different S3 bucket than the rest of the files, when this repository was first created.  These were previously referenced from the Registry DB as absolute URLs, which are going to break when the data moves to its new home at SLAC.