Add additional datastores for dp02
In the Butler configuration used for Butler server and DirectButler, add new relative roots that can be used to reference the DP02 raw image files and files from DP01.

Raw images for DP02 and datasets imported from DP01 were stored in a different S3 buckets than the rest of the files, when this repository was first created.  These were previously referenced from the Registry DB as absolute URLs, which are going to break when the data moves to its new home at SLAC.