Clean up Gafaelfawr Helm chart
Use a different method for handling Gafaelfawr schema updates: add
the hook metadata to the required supporting resources rather than
creating new ones. We tested this approach with vo-cutouts and it
worked well, so adopt it here as well. This results in replacing
instead of updating the resources on the next sync, but it avoids
having duplicate resources that have to be manually deleted.

Remove the -config and -secret endings on the names of the Gafaelfawr
VaultSecret and ConfigMap resources, adopting our current conventions.
The endings duplicate information already available in Kubernetes in
other ways.