Allow excluding applications from secrets sync
Sometimes we'll want to perform a secrets sync without affecting
a particular application, such as when we're doing experiments with
that application or when we want to run `--delete` but not affect
that application. Add an `--exclude` flag to both secrets sync and
secrets audit.

This currently excludes the application entirely, as if it doesn't
exist, and therefore may interfere with copy directives. We can do
something more sophisticated later if that ends up being a problem.