Bump Wobbly to 0.2.0, add maintenance CronJob
Upgrade Wobbly to 0.2.0 and add a maintenance `CronJob` that runs
`wobbly expire` to delete all jobs that have passed their expiration
times. Use the same requests and limits as the main Wobbly pod for
simplicity.

Delete a stray environment variable from the main Wobbly deployment
that was left over from cut and paste.