Add remaining Nublado config parameters
We've added several new configuration parameters to Nublado in
support of container flexibility and to add the ability to change
how `/tmp` is managed, but they weren't documented in the Phalanx
`values.yaml` file. Add the new settings with their default values.

Move the configuration of resources for the proxy pod to where it
will take effect (previously it was ignored).

Delete some redundant commentary. Shorten some defaults to try to
make the web formatting a bit better. Delete a commented-out block
that isn't supported by Zero to JupyterHub.