Use alpine version of sql proxy for schema updates
This was part of the Gafaelfawr implementation of the schema-update job,
but in the port to times square I missed the additional
schemaUpdateTagSuffix variable. By using an alpine-based image the
container gets a /bin/sh shell and therefore can run with the while loop
lifecycle logic.

This resolves a unable to start container process: exec: "/bin/sh": stat
/bin/sh: no such file or directory: unknown error.