DM-48733: times-square - change path filtering for the templating ingress
There is a [new template rendering
endpoint](https://github.com/lsst-sqre/times-square/pull/92) in times
square that finds a notebook template to render based on the GitHub URL
path. We want to [use this new
endpoint](https://github.com/lsst-sqre/rsp-jupyter-extensions/pull/42)
in our Jupyter extension to generate a notebook from a portal query so
that we can further restrict this usage to only templates from certain
GitHub repos, in this case, the lsst-sqre/nublado-seeds repo.