Use sidecar containers for Cloud SQL
Run the Cloud SQL Auth Proxy as a proper sidecar container using
the new feature in Kubernetes 1.29. This avoids having to do
coordination between the main container and the Cloud SQL Auth Proxy
container to shut down the latter for jobs, and also ensures the
correct startup order and shutdown semantics.