DM-49030: mobu - sentry release
Set a `SENTRY_RELEASE` env var to the mobu deployment which will tag
every Sentry event with a release, which will let us see which errors
show up in which releases.

We'll set this to the image tag, which is (I think) the best indicator
that we have for what a release is available to the template.