Move DP02 data product Butler root
The weekly 2022_40 stack version used for some tutorials does not accept a trailing directory name in a datastore root with Weka S3, because it enforces that the root `exists()` which it does not in S3 because it is a directory.

We are moving the directory on the filesystem so it points at the root.