Add the rebase_checker workflow
This reusable workflow checks if the main branch has been merged into a
PR's branch. This commonly happens when folks update their branch with
the GitHub UI. The downside of this approach is that it makes Git
history really hard to understand in a PR because commits from main are
intermixed with new commits. This workflows encourages people to use the
rebases, with a link to the Developer Guide.