Remove stand-alone Gafaelfawr Cloud SQL Proxy
Previously, Gafaelfawr installed a stand-alone Cloud SQL Proxy
service with its own `Service` and `NetworkPolicy` for the operator
pod and the cron jobs to use. Now that there is a better way to
handle sidecar containers for cron jobs and it's been moved into a
helper macro, switch all containers to using a sidecar and remove
the separate service.