Fix default scopes for user servers
We want to grant servers the ability to delete themselves since
that's how the logout menu works. This previously was done by giving
the server full `self` access, but that provides access to all sorts
of other things we don't want the server to be doing, such as creating
tokens for the user. Restrict that to only the permissions normally
given to servers, plus the ability to delete servers.