Bump limits for JupyterHub proxy
Increase the limit for the JupyterHub proxy to a full CPU and double
the memory limit. The current requests are accurate for normal usage
right now, but under load testing the JupyterHub proxy ran out of
memory and was OOM-killed, and we're seeing warnings about it hitting
its CPU limits.