Fix secrets sync with modify and delete
Previously, if `phalanx secrets sync --delete` needed to modify a
secret key's value and also delete a different secret key, it would
inadvertantly undo the first operation when doing the second due to
an out-of-date in-memory data structure. Fix this by updating the
in-memory copy of the Vault secret when secrets are added or modified
so that a subsequent delete uses the correct values.