mpsky: initial functional deployment
The image has now been pointed to ghcr.io/mjuric/mpsky-daily, and the
starting command updated accordingly.  At the moment, the image hardcodes
some of the input files which will soon be moved to a bucket.

This service needs to be accessible only internally from within USDF.  On
KTL's recommendation, we've therefore dropped the Ingress and changed the
service type to a LoadBalancer instead, using an IP from sdf-services pool.

See discussion at:
https://rubin-obs.slack.com/archives/C07Q45N7KHV/p1742520202629099